+++ src/teams/teams.controller.ts
@@ -17,6 +17,7 @@ import {
     VoyageTeamResponse,
 } from "./teams.response";
 import {
+    ForbiddenErrorResponse,
     NotFoundErrorResponse,
     UnauthorizedErrorResponse,
 } from "../global/responses/errors";
@@ -38,6 +39,16 @@ export class TeamsController {
         description: "Successfully gets all voyage teams",
         type: VoyageTeamResponse,
         isArray: true,
+    })
+    @ApiResponse({
+        status: HttpStatus.UNAUTHORIZED,
+        description: "unauthorized access - user is not logged in",
+        type: UnauthorizedErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.FORBIDDEN,
+        description: "forbidden - user does not have the required permission",
+        type: ForbiddenErrorResponse,
     })
     @CheckAbilities({ action: Action.Manage, subject: "all" })
     @Get()
@@ -69,7 +80,17 @@ export class TeamsController {
         required: true,
         example: 1,
     })
-    @CheckAbilities({ action: Action.Manage, subject: "VoyageTeam" })
+    @ApiResponse({
+        status: HttpStatus.UNAUTHORIZED,
+        description: "unauthorized access - user is not logged in",
+        type: UnauthorizedErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.FORBIDDEN,
+        description: "forbidden - user does not have the required permission",
+        type: ForbiddenErrorResponse,
+    })
+    @CheckAbilities({ action: Action.Manage, subject: "all" })
     @Get("voyages/:voyageId")
     findTeamsByVoyageId(@Param("voyageId", ParseIntPipe) voyageId: number) {
         return this.teamsService.findTeamsByVoyageId(voyageId);